output_connectors/gcp_bigquery/requirements.txt
google-cloud-bigquery==1.22.0
google-cloud-core==1.0.3
ubiops-connector==1.0.2
